gift_registry/requirements.txt

django~=3.1.1
django-extensions~=3.0.8
djangorestframework~=3.11.1
python-dotenv~=0.9.1

